requirements.txt
appdirs==1.4.3
Django==1.10.5
djangorestframework==3.6.2
djangorestframework-jwt==1.9.0
packaging==16.8
PyJWT==1.4.2
pyparsing==2.2.0
six==1.10.0
gunicorn==19.7.1
